Simple script that extracts songs from Beat Saber.
Edit the .py file and change the file paths.  
One is the Beat Saber installation path, the other is where you want the songs to go.

Run the Run.bat and it will execute the script, then visit the destination path you input, and all your songs will be there.

Songs are stored in .ogg format.